Move completeStaleSessions to SyncableObject.completeStale
