+++ Sources/Scout/Core/Sync/SyncableObject.swift
@@ -24,6 +24,10 @@ class SyncableObject: IDObject {
     @NSManaged var isSynced: Bool
     @NSManaged var syncAttempts: Int
 
+    static var stalePredicate: NSPredicate {
+        NSPredicate(format: "endDate == nil AND launchID != %@", IDs.launch as CVarArg)
+    }
+
     static var pendingPredicate: NSPredicate {
         NSPredicate(format: "isSynced == false AND syncAttempts <= %d", maxSyncAttempts)
     }
